The game is now saved and loaded by pressing 's' and 'l' buttons accordingly. Well! One could say that the prototype is ready to be extended to a real game, and I'll totally agree.

+++ Assets/Scripts/Building.cs
@@ -18,7 +18,7 @@ public class BuildingInfo
 {
     public int id;
     public int connectedGridId;
-    public float level = 0;
+    public int level = 0;
     public float yRotation = 0;
 }
